android-sonet update

+++ src/com/piusvelte/sonet/SonetService.java
@@ -126,7 +126,6 @@ public class SonetService extends Service implements Runnable {
 	@Override
 	public void onStart(Intent intent, int startId) {
 		super.onStart(intent, startId);
-		Log.v(TAG,"onStart");
 		if (intent != null) {
 			if ((intent.getAction() != null) && (!intent.getAction().equals(ACTION_REFRESH))) SonetService.updateWidgets(new int[]{Integer.parseInt(intent.getAction())});
 			else if (intent.hasExtra(AppWidgetManager.EXTRA_APPWIDGET_IDS)) SonetService.updateWidgets(intent.getIntArrayExtra(AppWidgetManager.EXTRA_APPWIDGET_IDS));
@@ -276,7 +275,6 @@ public class SonetService extends Service implements Runnable {
 		long now = new Date().getTime();
 		while (!accounts.isAfterLast()) {
 			int service = accounts.getInt(iservice);
-			Log.v(TAG,"service: "+service);
 			switch (service) {
 			case TWITTER:
 				String status_url = "http://twitter.com/%s/status/%s";
@@ -412,8 +410,7 @@ public class SonetService extends Service implements Runnable {
 		AlarmManager alarmManager = (AlarmManager) getSystemService(Context.ALARM_SERVICE);
 		Boolean hasbuttons,
 		time24hr,
-		hasAccount = true,
-		scrollable = false;
+		hasAccount = true;
 		int interval,
 		buttons_bg_color,
 		buttons_color,
@@ -425,8 +422,7 @@ public class SonetService extends Service implements Runnable {
 		while (updatesQueued()) {
 			int appWidgetId = getNextUpdate();
 			alarmManager.cancel(PendingIntent.getService(this, 0, new Intent(this, SonetService.class).setAction(Integer.toString(appWidgetId)), 0));
-			Log.v(TAG,"appWidgetId: "+appWidgetId);
-			Cursor settings = db.rawQuery("select " + _ID + "," + INTERVAL + "," + HASBUTTONS + ","	+ BUTTONS_BG_COLOR + "," + BUTTONS_COLOR + "," + MESSAGES_BG_COLOR + "," + MESSAGES_COLOR + "," + FRIEND_COLOR + "," + CREATED_COLOR + "," + TIME24HR + "," + SCROLLABLE + " from " + TABLE_WIDGETS + " where " + WIDGET + "=" + appWidgetId, null);
+			Cursor settings = db.rawQuery("select " + _ID + "," + INTERVAL + "," + HASBUTTONS + ","	+ BUTTONS_BG_COLOR + "," + BUTTONS_COLOR + "," + MESSAGES_BG_COLOR + "," + MESSAGES_COLOR + "," + FRIEND_COLOR + "," + CREATED_COLOR + "," + TIME24HR + " from " + TABLE_WIDGETS + " where " + WIDGET + "=" + appWidgetId, null);
 			if (settings.getCount() > 0) {
 				settings.moveToFirst();
 				interval = settings.getInt(settings.getColumnIndex(INTERVAL));
@@ -438,7 +434,6 @@ public class SonetService extends Service implements Runnable {
 				friend_color = settings.getInt(settings.getColumnIndex(FRIEND_COLOR));
 				created_color = settings.getInt(settings.getColumnIndex(CREATED_COLOR));
 				time24hr = settings.getInt(settings.getColumnIndex(TIME24HR)) == 1;
-				scrollable = settings.getInt(settings.getColumnIndex(SCROLLABLE)) == 1;
 			} else {
 				// upgrade, moving settings from sharedpreferences to db
 				if (sp == null) sp = (SharedPreferences) getSharedPreferences(getString(R.string.key_preferences), SonetService.MODE_PRIVATE);
@@ -499,7 +494,6 @@ public class SonetService extends Service implements Runnable {
 							c.moveToNext();
 						}
 					} else {
-						Log.v(TAG,"no accounts");
 						hasAccount = false;
 						// alert user: no accounts setup
 						statuses = new ArrayList<StatusItem>();
@@ -514,7 +508,6 @@ public class SonetService extends Service implements Runnable {
 				statuses = new ArrayList<StatusItem>();
 				Cursor cached = db.rawQuery("select " + _ID + "," + CREATED + "," + LINK + "," + FRIEND + "," + PROFILE + "," + MESSAGE + "," + SERVICE + "," + CREATEDTEXT + " from " + TABLE_STATUSES + " where " + WIDGET + "=" + appWidgetId + " order by " + CREATED + " desc", null);
 				if (cached.getCount() > 0) {
-					Log.v(TAG,"load from cache");
 					cached.moveToFirst();
 					int created = cached.getColumnIndex(CREATED),
 					link = cached.getColumnIndex(LINK),
@@ -535,12 +528,60 @@ public class SonetService extends Service implements Runnable {
 					}
 				} else statuses.add(new StatusItem(0, null, null, null, "no connection", 0, null)); // alert user: no connection, no cache
 			}
-			if (hasConnection) {
-				Log.v(TAG,"updating cache");
-				// clear the cache
-				db.execSQL("delete from " + TABLE_STATUSES + ";");
-				// update the cache
-				for  (StatusItem item : statuses) {
+			if (hasConnection) db.execSQL("delete from " + TABLE_STATUSES + ";"); // clear the cache
+			// race condition when finished configuring, the service starts.
+			// meanwhile, the launcher broadcasts READY and the listview is created. it's at this point that the widget is marked scrollable
+			// this run finishes after the listview is created, but is not flagged as scrollable and replaces the listview with the regular widget
+			boolean scrollable = false;
+			settings = db.rawQuery("select " + _ID + "," + SCROLLABLE + " from " + TABLE_WIDGETS + " where " + WIDGET + "=" + appWidgetId, null);
+			if (settings.getCount() > 0) {
+				settings.moveToFirst();
+				scrollable = settings.getInt(settings.getColumnIndex(SCROLLABLE)) == 1;
+			}
+			// Push update for this widget to the home screen
+			// set messages background
+			Bitmap messages_bg = Bitmap.createBitmap(1, 1, Config.ARGB_8888);
+			Canvas messages_bg_canvas = new Canvas(messages_bg);
+			messages_bg_canvas.drawColor(messages_bg_color);
+			int[] map_item = {R.id.item0, R.id.item1, R.id.item2, R.id.item3, R.id.item4, R.id.item5, R.id.item6},
+			map_profile = {R.id.profile0, R.id.profile1, R.id.profile2, R.id.profile3, R.id.profile4, R.id.profile5, R.id.profile6},
+			map_message = {R.id.message0, R.id.message1, R.id.message2, R.id.message3, R.id.message4, R.id.message5, R.id.message6},
+			map_screenname = {R.id.friend0, R.id.friend1, R.id.friend2, R.id.friend3, R.id.friend4, R.id.friend5, R.id.friend6},
+			map_created = {R.id.created0, R.id.created1, R.id.created2, R.id.created3, R.id.created4, R.id.created5, R.id.created6};
+			RemoteViews views = new RemoteViews(getPackageName(), hasbuttons ? R.layout.widget : R.layout.widget_nobuttons);
+			if (hasbuttons) {
+				Bitmap buttons_bg = Bitmap.createBitmap(1, 1, Config.ARGB_8888);
+				Canvas buttons_bg_canvas = new Canvas(buttons_bg);
+				buttons_bg_canvas.drawColor(buttons_bg_color);
+				views.setImageViewBitmap(R.id.buttons_bg, buttons_bg);
+				views.setTextColor(R.id.head_spacer, buttons_bg_color);
+				views.setOnClickPendingIntent(R.id.button_post, PendingIntent.getActivity(this, 0, new Intent(this, PostDialog.class), 0));
+				views.setTextColor(R.id.button_post, buttons_color);
+				views.setOnClickPendingIntent(R.id.button_configure, PendingIntent.getActivity(this, 0, new Intent(this, UI.class).setAction(Integer.toString(appWidgetId)), 0));
+				views.setTextColor(R.id.button_post, buttons_color);
+				views.setOnClickPendingIntent(R.id.button_refresh, PendingIntent.getService(this, 0, new Intent(this, SonetService.class).setAction(Integer.toString(appWidgetId)), 0));
+				views.setTextColor(R.id.button_post, buttons_color);
+			}
+			views.setImageViewBitmap(R.id.messages_bg, messages_bg);
+			int count_status = 0;
+			for  (StatusItem item : statuses) {
+				if (!scrollable && (count_status < map_item.length)) {
+					views.setTextViewText(map_message[count_status], item.message);
+					views.setTextColor(map_message[count_status], messages_color);
+					if (item.friend != null) {
+						// if no buttons, use StatusDialog.java with options for Config and Refresh
+						if (hasbuttons) views.setOnClickPendingIntent(map_item[count_status], PendingIntent.getActivity(this, 0, new Intent(Intent.ACTION_VIEW, Uri.parse(item.link)), 0));
+						else views.setOnClickPendingIntent(map_item[count_status], PendingIntent.getActivity(this, 0, new Intent(this, StatusDialog.class).setAction(appWidgetId+"`"+item.service+"`"+item.link), 0));
+						views.setTextViewText(map_screenname[count_status], item.friend);
+						views.setTextColor(map_screenname[count_status], friend_color);
+						views.setTextViewText(map_created[count_status], item.createdText);
+						views.setTextColor(map_created[count_status], created_color);
+						views.setImageViewBitmap(map_profile[count_status], BitmapFactory.decodeByteArray(item.profile, 0, item.profile.length));						
+					}
+					count_status++;
+				}
+				if (hasConnection) {
+					// update the cache
 					ContentValues values = new ContentValues();
 					values.put(CREATED, item.created);
 					values.put(LINK, item.link);
@@ -553,60 +594,8 @@ public class SonetService extends Service implements Runnable {
 					db.insert(TABLE_STATUSES, _ID, values);
 				}
 			}
-			// PROBLEM, race condition: when finished configuring, the service starts.
-			// meanwhile, the launcher broadcasts READY and the listview is created. it's at this point that the widget is marked scrollable
-			// this run finishes after the listview is created, but is not flagged as scrollable and replaces the listview with the regular widget
-			if (scrollable) {
-				Log.v(TAG,"notify scrollable");
-				SonetProvider.notifyDatabaseModification(this, appWidgetId);
-			} else {
-				Log.v(TAG, "is not scrollable, push update");
-				// Push update for this widget to the home screen
-				// set messages background
-				Bitmap messages_bg = Bitmap.createBitmap(1, 1, Config.ARGB_8888);
-				Canvas messages_bg_canvas = new Canvas(messages_bg);
-				messages_bg_canvas.drawColor(messages_bg_color);
-				int[] map_item = {R.id.item0, R.id.item1, R.id.item2, R.id.item3, R.id.item4, R.id.item5, R.id.item6},
-				map_profile = {R.id.profile0, R.id.profile1, R.id.profile2, R.id.profile3, R.id.profile4, R.id.profile5, R.id.profile6},
-				map_message = {R.id.message0, R.id.message1, R.id.message2, R.id.message3, R.id.message4, R.id.message5, R.id.message6},
-				map_screenname = {R.id.friend0, R.id.friend1, R.id.friend2, R.id.friend3, R.id.friend4, R.id.friend5, R.id.friend6},
-				map_created = {R.id.created0, R.id.created1, R.id.created2, R.id.created3, R.id.created4, R.id.created5, R.id.created6};
-				RemoteViews views = new RemoteViews(getPackageName(), hasbuttons ? R.layout.widget : R.layout.widget_nobuttons);
-				if (hasbuttons) {
-					Bitmap buttons_bg = Bitmap.createBitmap(1, 1, Config.ARGB_8888);
-					Canvas buttons_bg_canvas = new Canvas(buttons_bg);
-					buttons_bg_canvas.drawColor(buttons_bg_color);
-					views.setImageViewBitmap(R.id.buttons_bg, buttons_bg);
-					views.setTextColor(R.id.head_spacer, buttons_bg_color);
-					views.setOnClickPendingIntent(R.id.button_post, PendingIntent.getActivity(this, 0, new Intent(this, PostDialog.class), 0));
-					views.setTextColor(R.id.button_post, buttons_color);
-					views.setOnClickPendingIntent(R.id.button_configure, PendingIntent.getActivity(this, 0, new Intent(this, UI.class).setAction(Integer.toString(appWidgetId)), 0));
-					views.setTextColor(R.id.button_post, buttons_color);
-					views.setOnClickPendingIntent(R.id.button_refresh, PendingIntent.getService(this, 0, new Intent(this, SonetService.class).setAction(Integer.toString(appWidgetId)), 0));
-					views.setTextColor(R.id.button_post, buttons_color);
-				}
-				views.setImageViewBitmap(R.id.messages_bg, messages_bg);
-				int count_status = 0;
-				for  (StatusItem item : statuses) {
-					if (count_status < map_item.length) {
-						views.setTextViewText(map_message[count_status], item.message);
-						views.setTextColor(map_message[count_status], messages_color);
-						if (item.friend != null) {
-							// if no buttons, use StatusDialog.java with options for Config and Refresh
-							if (hasbuttons) views.setOnClickPendingIntent(map_item[count_status], PendingIntent.getActivity(this, 0, new Intent(Intent.ACTION_VIEW, Uri.parse(item.link)), 0));
-							else views.setOnClickPendingIntent(map_item[count_status], PendingIntent.getActivity(this, 0, new Intent(this, StatusDialog.class).setAction(appWidgetId+"`"+item.service+"`"+item.link), 0));
-							views.setTextViewText(map_screenname[count_status], item.friend);
-							views.setTextColor(map_screenname[count_status], friend_color);
-							views.setTextViewText(map_created[count_status], item.createdText);
-							views.setTextColor(map_created[count_status], created_color);
-							views.setImageViewBitmap(map_profile[count_status], BitmapFactory.decodeByteArray(item.profile, 0, item.profile.length));						
-						}
-						count_status++;
-					}
-				}
-				Log.v(TAG,"updateAppWidget");
-				appWidgetManager.updateAppWidget(appWidgetId, views);
-			}
+			appWidgetManager.updateAppWidget(appWidgetId, views);
+			if (scrollable) SonetWidget.buildScrollable(this, appWidgetId); // replace with scrollable widget
 			if (hasAccount && (interval > 0)) alarmManager.set(AlarmManager.RTC, System.currentTimeMillis() + interval, PendingIntent.getService(this, 0, new Intent(this, SonetService.class).setAction(Integer.toString(appWidgetId)), 0));
 		}
 		db.close();
